Added throttling to secure server and db

+++ src/websocket/websock.js
@@ -8,6 +8,7 @@ const query= require('../db/query');
 const server = createServer(app);
 const io = new Server(server);
 
+const rateMap = {};
 
 io.use((socket, next) => {
   try {
@@ -32,6 +33,7 @@ io.use((socket, next) => {
 
 io.on('connection', (socket) => {
     console.log('A user connected.');
+    rateMap[socket.user_id] = 0;
 
     if (socket.user_id) {
         const userRoom = `user_${socket.user_id}`;
@@ -40,25 +42,48 @@ io.on('connection', (socket) => {
     }
 
     socket.on('added', async (newItem, targ) => {
+        if(rateMap[socket.user_id] > 1) {
+            rateMap[socket.user_id]++;
+            socket.emit('rate-exceed', {msg:"Slow down, I don't own Google"});
+            return;
+        }
+        rateMap[socket.user_id]++;
         const newItemId = await query.addItem(newItem, targ, socket.user_id);
         console.log(`Item added: ${newItemId}-${newItem} by ${targ}`);
         io.to(`user_${socket.user_id}`).emit('addNew', newItemId, newItem, targ);
     });
 
     socket.on('update', async (newItemId, newItem, targ) => {
+        if(rateMap[socket.user_id] > 1) {
+            rateMap[socket.user_id]++;
+            socket.emit('rate-exceed', {msg:"Slow down, I don't own Google"});
+            return;
+        }
+        rateMap[socket.user_id]++;
         console.log(`Item updated: ${newItemId}-${newItem} by ${targ}`);
         await query.updateList(newItemId, newItem, targ);
         io.to(`user_${socket.user_id}`).emit('upd', newItemId, newItem, targ);
     });
 
     socket.on('deleted', async (deleteId) => {
+        if(rateMap[socket.user_id] > 1) {
+            rateMap[socket.user_id]++;
+            socket.emit('rate-exceed', {msg:"Slow down, I don't own Google"});
+            return;
+        }
+        rateMap[socket.user_id]++;
         console.log(`Item will be deleted: ${deleteId}`);
         await query.deleteItem(deleteId);
         io.to(`user_${socket.user_id}`).emit('del', deleteId);
     });
 
+    setTimeout(() => {
+        rateMap[socket.user_id]=0;
+    }, 200);
+
     socket.on('disconnect', () => {
         console.log(`A user disconnected.`);
+        delete rateMap[socket.user_id];
     });
 });
